Fix a bug of blank line in ip-list from url.
If a blank line is in ip-list from url, the module of earlier version will match all ip,
but a blank line is invalid in directives defined by mod_authz_host, so it should not be accepted in mod_auth_remote.

The function dealing with ip-match problem used by mod_authz_host will cause blank line to match all ip,
but a directive with blank line argument will not be accpeted by server.
Consider such directive "allow from      ", it will cause a start-up error of apache, so this would not be a bug in mod_authz_host.

But in mod_auth_remote, server does not know there is a blank line in ip-list from url until it has started,
so this function will cause a bug in mod_auth_remote.